refactor: organize ui files
