+++ lib/ui/features/auth/views/change_password/change_password_success_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_base/ui/components/box_spacer.dart';
 import 'package:flutter_base/ui/components/buttons/button_primary.dart';
 import 'package:flutter_base/ui/components/flutter_base_app_bar.dart';
-import 'package:flutter_base/ui/components/scaffold_bottom_sheet.dart';
+import 'package:flutter_base/ui/components/sheets/scaffold_bottom_sheet.dart';
 import 'package:flutter_base/ui/components/text/high_text.dart';
 import 'package:flutter_base/ui/components/views/column_scroll_view.dart';
 import 'package:flutter_base/ui/extensions/media_query.dart';
